添加完善简单光照支持，实现rayCast下的光照可见区域，相关代码修改

+++ src/utils/vec2.h
@@ -13,7 +13,7 @@ public:
 
 	Vec2<T> operator+(const Vec2<T>& other)const;
 	Vec2<T> operator-(const Vec2<T>& other)const;
-	Vec2<T> operator*(int factor)const;
+	Vec2<T> operator*(float factor)const;
 	Vec2<T> operator/(int divisior)const;
 
 	Vec2<T>&operator+= (const Vec2<T>& other);
